services/mcp-odoo/pyproject.toml
[project]
name = "mcp-odoo"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.2.0",
  "uvicorn>=0.30.0",
  "starlette>=0.37.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.0",
  "requests>=2.32.5",
]
